requirements.txt
python-telegram-bot[job-queue]==21.*
aiohttp>=3.9,<4
